Refactor: 마이페이지 코드 리팩토링
