+++ src/components/domain/admin/mypage/ChangeBox.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { Button, ShadowBox, Txt } from '@/components/atoms';
