+++ src/components/domain/admin/mypage/index.ts
@@ -1,0 +1,2 @@
+export { default as Profile } from './Profile';
+export { default as ChangeBox } from './ChangeBox';
